+++ tests/cli.rs
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 use assert_cmd::Command;
-use std::{env::set_current_dir, fs::File};
+use std::fs::File;
 use tempfile::TempDir;
 
 #[test]
@@ -90,7 +90,6 @@ fn basic_compilation_from_toml() {
 
     let solana_test = tmp.path().join("solana_test");
 
-    println!("path:  =>> {}", solana_test.display());
     //solang new --target solana
     new_cmd
         .arg("new")
@@ -98,14 +97,11 @@ fn basic_compilation_from_toml() {
         .args(["--target", "solana"])
         .assert()
         .success();
-
-    let flipper_file = File::open(solana_test.join("flipper.sol")).expect("should exist");
-    let config_file = File::open(solana_test.join("Solang.toml")).expect("should exist");
+    File::open(solana_test.join("flipper.sol")).expect("should exist");
+    File::open(solana_test.join("Solang.toml")).expect("should exist");
 
     // compile flipper using config file
     let mut compile_cmd = Command::cargo_bin("solang").unwrap();
-
-    //assert!(set_current_dir(solana_test).is_ok());
 
     compile_cmd
         .args(["compile", "--configuration-file"])
@@ -114,7 +110,7 @@ fn basic_compilation_from_toml() {
         .success();
 
     let substrate_test = tmp.path().join("substrate_test");
-    let new_cmd = Command::cargo_bin("solang")
+    let _new_cmd = Command::cargo_bin("solang")
         .unwrap()
         .arg("new")
         .arg(substrate_test.clone())
@@ -122,9 +118,5 @@ fn basic_compilation_from_toml() {
         .assert()
         .success();
 
-    //assert!(set_current_dir(substrate_test).is_ok());
-
     compile_cmd.current_dir(substrate_test).assert().success();
 }
-
-fn incorrect_toml_error_handling() {}
